chore: remove old BRP-related components and update residents data fetching
