--- src/app/pages/addresses/peopleNewBrp/PeopleNewBrp.tsx
@@ -1,39 +1,0 @@
-
-import { useParams } from "react-router-dom"
-import isValidUrlParamBAGId from "app/routing/utils/isValidUrlParamBAGId"
-import DefaultLayout from "app/components/layouts/DefaultLayout/DefaultLayout"
-import Row from "app/components/layouts/Grid/Row"
-import Column from "app/components/layouts/Grid/Column"
-import DetailHeader from "app/components/shared/DetailHeader/DetailHeader"
-import ResidentsOverviewNew from "app/components/addresses/ResidentsOverviewNew/ResidentsOverviewNew"
-import PageHeading from "app/components/shared/PageHeading/PageHeading"
-import NotFoundPage from "app/pages/errors/NotFoundPage"
-
-type Props = {
-  bagId: string
-}
-
-const PeopleNewBrp: React.FC = () => {
-  const { bagId } = useParams<Props>()
-  return (
-    isValidUrlParamBAGId(bagId) ? (
-      <DefaultLayout>
-        <Row>
-          <Column spanLarge={ 50 }>
-            <PageHeading />
-          </Column>
-          <Column spanLarge={ 50 }>
-            <DetailHeader bagId={ bagId } />
-          </Column>
-        </Row>
-        <Row>
-          <Column spanLarge={ 50 }>
-            <ResidentsOverviewNew bagId={ bagId } />
-          </Column>
-        </Row>
-      </DefaultLayout>
-    ) : <NotFoundPage />
-  )
-}
-
-export default PeopleNewBrp
