+++ src/app/state/rest/index.ts
@@ -36,7 +36,6 @@ export * from "./permissions"
 export * from "./processes"
 export * from "./reasons"
 export * from "./residents"
-export * from "./residentsNew"
 export * from "./roles"
 export * from "./schedules"
 export * from "./tasks"
